add docker-compose-ed project infrastructure
Testing this project is a little tricky without an OAuth2 endpoint so
we update the docker-compose infrastructure to create as much
infrastructure as possible, including the hydra OAuth2 server and a
testing consent app.

The consent app simply grants whatever scopes were requested and lets
you set the username associated with the token.

Add some simple helper scripts for creating clients and tokens for use
with the OAuth2 endpoint.