implement lookupapi application
Add lookupapi, an application which implements a proxy REST-y API to the
anonymous lookup service. It is authenticated by means of an OAuth2
token and set of required scopes for that token. Tokens are validated by
a RFC7662[1] introspection endpoint.

Documentation for the API is by means of an OpenAPI document which is
rendered by a Swagger UI instance at /ui.

The API is read-only.

[1] https://tools.ietf.org/html/rfc7662