generalise person detail view to let scheme be a parameter
Lookup, in principle, supports more than the crsid scheme. We will want
to add a "test" scheme for test users so we will need to support at
least two schemes. Add the required plumbing to turn the currently
hard-wired people/crsid/<crsid> endpoint into
people/<scheme>/<identifier>.

We currently pass the scheme directly to lookup since it is the ultimate
arbiter of what schemes are valid.